Config files' creation, spelling errors, and defaul prefix fixes
Now checking the existance of config.json, servers.json, and .env. If the .env file doesn't exist the program has to be restarted after adding the token
Fixed spelling errors
The bot's status now updates based on the default prefix decleared in config.json. The help command now always works with the default prefix instead of !.